requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
azure-cognitiveservices-speech==1.29.0
boto3==1.26.158
botocore==1.29.158
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
decorator==4.4.2
discord==2.3.0
discord.py==2.3.0
ffmpeg==1.4
ffmpeg-python==0.2.0
frozenlist==1.3.3
future==0.18.3
gstreamer-player==1.1.2
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
jmespath==1.0.1
moviepy==1.0.3
multidict==6.0.4
mutagen==1.46.0
mysql-connector-python==8.0.33
numpy==1.25.0
openai==0.27.8
Pillow==9.5.0
proglog==0.1.10
proto-plus==1.22.2
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
regex==2023.6.3
requests==2.31.0
rsa==4.9
s3transfer==0.6.1
six==1.16.0
tqdm==4.65.0
urllib3==1.26.16
yarl==1.9.2
